requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
feedparser==6.0.11
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
